moved file query utilities to their own file
